--- Assets/RefreshButton.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7018ec64e083a4239bc0814c6a991781
-timeCreated: 1457998288
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
