+++ Assets/Resources/Prefab/Incantation.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7789e4c9bce374058a355ae5f27d3b4a
+timeCreated: 1458056975
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
