+++ Assets/Scripts/General/GameManager/GameManagerScript.cs
@@ -2,13 +2,14 @@ using UnityEngine;
 using UnityEngine.UI;
 using System.Collections;
 
-public class GameManagerScript : MonoBehaviour, IPlayerInstantiationController, IEggInstantiationController, ILevelLoader {
+public class GameManagerScript : MonoBehaviour, IPlayerInstantiationController, IEggInstantiationController, IIncantationInstantiationController, ILevelLoader {
 	// Static instance
 	public static GameManagerScript 		instance;
 
 	// Prefabs
 	private GameObject						_playerPrefab;
 	private GameObject						_eggPrefab;
+	private GameObject						_incantationPrefab;
 
 	// Managers
 	public GridController					gridController { get; private set; }
@@ -16,7 +17,7 @@ public class GameManagerScript : MonoBehaviour, IPlayerInstantiationController, 
 	public TimeManager						timeManager { get; private set; }
 	public TeamManager						teamManager { get; private set; }
 	public InputManager 					inputManager { get; private set; }
-	public PlayerManagerScript				playerManager { get; private set; }
+	public PlayerManager				playerManager { get; private set; }
 	public DebugManager						debugManager { get; private set; }	
 	public ServerCommands					commandsManager { get; private set; }
 
@@ -30,13 +31,14 @@ public class GameManagerScript : MonoBehaviour, IPlayerInstantiationController, 
 		DontDestroyOnLoad (this);
 		_playerPrefab = Resources.Load ("Prefab/Player") as GameObject;
 		_eggPrefab = Resources.Load ("Prefab/Egg(Teleporter)") as GameObject;
+		_incantationPrefab = Resources.Load ("Prefab/Incantation") as GameObject;
 
 		gridController = GetComponentInChildren<GridScript>().controller;
 		qualityManager = new QualityManager();
 		timeManager = new TimeManager();
 		teamManager = new TeamManager();
 		inputManager = new InputManager();
-		playerManager = new PlayerManagerScript(gridController, teamManager, this, this);
+		playerManager = new PlayerManager(gridController, teamManager, this, this);
 		debugManager =  new DebugManager();
 		commandsManager = new ServerCommands(gridController, teamManager, playerManager, timeManager, debugManager, this);
 	}
@@ -44,14 +46,22 @@ public class GameManagerScript : MonoBehaviour, IPlayerInstantiationController, 
 	// Prefab instantiation
 	public EggController InstantiateEgg ()
 	{
-		GameObject clone = Instantiate (_eggPrefab);
+		var clone = Instantiate (_eggPrefab);
 		return clone.GetComponent<EggScript>().controller;
 	}
 	
 	public PlayerController InstantiatePlayer ()
 	{
-		GameObject clone = Instantiate (_playerPrefab);
+		var clone = Instantiate (_playerPrefab);
 		return clone.GetComponent<PlayerScript>().controller;
+	}
+
+	public IncantationScript InstantiateIncantation (Vector3 position, Color color)
+	{
+		var clone = Instantiate (_incantationPrefab).GetComponent<IncantationScript>();
+		clone.transform.position = position;
+		clone.startColor = color;
+		return clone;
 	}
 
 	// GameOver event
@@ -104,4 +114,9 @@ public interface IEggInstantiationController
 public interface IPlayerInstantiationController
 {
 	PlayerController InstantiatePlayer ();
+}
+
+public interface IIncantationInstantiationController
+{
+	IncantationScript InstantiateIncantation(Vector3 position, Color color);
 }
