+++ Assets/Scripts/General/PlayerManager/PlayerManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 95676c4095c594f5591d5847304b4140
+timeCreated: 1458060249
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
