--- Assets/Scripts/General/PlayerManager/PlayerManagerScript.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: f14ec40d686e44de4a8a1c3ec2600cab
-timeCreated: 1457627004
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
