+++ Assets/Scripts/Player/IncantationScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4d486ba7bd0f843f295779ed14ab23c7
+timeCreated: 1458057896
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
